% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_helpers.R
\name{compute_logFC}
\alias{compute_logFC}
\title{Compute Log-Fold Changes Between Conditions}
\usage{
compute_logFC(M, compStatus = NULL)
}
\arguments{
\item{M}{A numeric array of dimension \eqn{(p \times C \times V)},
where \eqn{p} is the number of genes, \eqn{C} is the number of
partitioning variable levels, and \eqn{V} is the number of conditions. Entries
represent estimated mean expression levels.}

\item{compStatus}{Character vector of the infection status levels to output
pairwise log fold changes. Default will return a list of all pairwise comparisons.}
}
\value{
A named list of data frames, one for each pairwise
comparison. Each data frame has \eqn{p} rows
(genes) and \eqn{C} columns (partitioning variable/cell types).
}
\description{
Computes log-fold change (logFC) contrasts between all pairs of
infection level stauses for each gene and partitioning variable leveles
(e.g., cell types), given estimated mean expression values.
}
\keyword{internal}
