% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.mu.null.R
\name{DE.mu.null}
\alias{DE.mu.null}
\title{Estimate Null Model Mean Expression}
\usage{
DE.mu.null(Y, W)
}
\arguments{
\item{Y}{A numeric matrix of dimension \eqn{(n \times p)}, where
\eqn{n} is the number of samples (cells) and \eqn{p} is the
number of features (genes).}

\item{W}{A numeric array of dimension \eqn{(n \times C \times V)},
where \eqn{C} is the number of cell types (or classes),
and \eqn{V} is the number of conditions (or contrasts).}
}
\value{
A numeric array of dimension \eqn{(p \times C \times V)}
containing the estimated mean expression values under the null model.
Specifically, the condition weights are replaced by their average
across \eqn{V}.
}
\description{
Computes the mean expression array under a null model, where
condition/variant-specific weights are averaged across conditions.
}
