% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.step1.R
\name{E.step1}
\alias{E.step1}
\title{E-step of Multiway Mixture Model}
\usage{
E.step1(Y, C.obs, V.obs, M, probs, sigma2)
}
\arguments{
\item{Y}{Expression matrix (cell x gene).}

\item{C.obs}{Observed condition labels (vector).}

\item{V.obs}{Observed infection status labels (vector).}

\item{M}{Mean array (gene x condition x infection status).}

\item{probs}{Mixing proportions (condition x infection status matrix).}

\item{sigma2}{Variance array (gene x condition x infection status).}
}
\value{
Weight tensor (cell x condition x infection status).
}
\description{
E-step of Multiway Mixture Model
}
