% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_e_step.R
\name{E_step}
\alias{E_step}
\title{E-step of Multiway Mixture Model}
\usage{
E_step(Y, c_obs, v_obs, M, probs, sigma2)
}
\arguments{
\item{Y}{Expression matrix (cell x gene).}

\item{c_obs}{Observed condition labels (vector).}

\item{v_obs}{Observed infection status labels (vector).}

\item{M}{Mean array (gene x condition x infection status).}

\item{probs}{Mixing proportions (condition x infection status matrix).}

\item{sigma2}{Variance array (gene x condition x infection status).}
}
\value{
Weight tensor (cell x condition x infection status).
}
\description{
E-step of Multiway Mixture Model
}
\keyword{internal}
