% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_set_inference.R
\name{getTestWeights}
\alias{getTestWeights}
\title{Compute Weights Matrix on Test Data}
\usage{
getTestWeights(mod_results, testData)
}
\arguments{
\item{mod_results}{A list containing model fitting results from \code{\link{fit_scDEcrypter}}.}

\item{testData}{A Seurat object or list containing test data.}
}
\value{
The W weight matrix (cells x cell types x viral types).
}
\description{
Computes the probability (weight) matrix for test data using the fitted model.
}
