% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M.step.probs.R
\name{M.step.probs}
\alias{M.step.probs}
\title{Update Mixing Proportions (M-step)}
\usage{
M.step.probs(Y, W)
}
\arguments{
\item{Y}{Numeric matrix of observed data with dimensions
\code{(n x p)}, where \code{n} is the number of samples (cells)
and \code{p} is the number of features (e.g., genes).}

\item{W}{Numeric 3D array of posterior weights (responsibilities)
with dimensions \code{(n x C x V)}, where \code{C} is the number
of conditions and \code{V} is the number of infection status}
}
\value{
A numeric matrix of updated mixing proportions with dimensions
\code{(C x V)}. Each entry corresponds to the expected proportion
of samples assigned to condition \code{c} and variant \code{v}.
}
\description{
Computes updated mixing proportions (class probabilities)
in the M-step of the multiway mixture model EM algorithm.
}
