% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M.step.variance.R
\name{M.step.variance}
\alias{M.step.variance}
\title{Update Variance Parameters (M-step)}
\usage{
M.step.variance(Y, W, M)
}
\arguments{
\item{Y}{Numeric matrix of observed data with dimensions
\code{(n x p)}, where \code{n} is the number of samples (cells)
and \code{p} is the number of features (e.g., genes).}

\item{W}{Numeric 3D array of posterior weights (responsibilities)
with dimensions \code{(n x C x V)}, where \code{C} is the number
of conditions and \code{V} is the number of infection status}

\item{M}{Numeric array of mean estimates with dimensions
\code{(p x C x V)}.}
}
\value{
A numeric array of updated variance estimates with dimensions
\code{(p x C x V)}.
}
\description{
Computes updated variance estimates \eqn{\sigma^2} in the M-step
of the multiway mixture model EM algorithm.
}
