% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_em.R
\name{run_penalized_em}
\alias{run_penalized_em}
\title{Run penalized generalized EM and retain every iterate}
\usage{
run_penalized_em(Y, c_obs, v_obs, M, sigma2, probs, lambda, max.iter, tol)
}
\description{
Run penalized generalized EM and retain every iterate
}
\keyword{internal}
