% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_helpers.R
\name{safe_weights}
\alias{safe_weights}
\title{Coerce Weights to a Three-Dimensional Array}
\usage{
safe_weights(weights)
}
\arguments{
\item{weights}{A matrix or three-dimensional array of weights.}
}
\value{
A three-dimensional array of weights.
}
\description{
Coerce Weights to a Three-Dimensional Array
}
\keyword{internal}
