% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDEcrypter-package.R
\docType{package}
\name{scDEcrypter-package}
\alias{scDEcrypter}
\alias{scDEcrypter-package}
\title{scDEcrypter: Viral Infection Analysis with Partially Labeled Single-Cell RNA Sequencing Data}
\description{
Fits a penalized Gaussian multiway mixture model to transformed single-cell RNA sequencing data with partially observed infection labels. Estimates infection-state probabilities and performs permutation-based differential expression testing.
}
\author{
\strong{Maintainer}: Rhonda Bacher \email{rbacher@ufl.edu}

Authors:
\itemize{
  \item Rhonda Bacher \email{rbacher@ufl.edu}
  \item Aaron Molstad
  \item Luer Zhong \email{luerzhong@ufl.edu}
}

}
\keyword{internal}
