% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_funcs.R, R/test_set_helpers.R
\name{select_by_threshold}
\alias{select_by_threshold}
\title{Select by thresholding}
\usage{
select_by_threshold(x, cutoff, f.labels)

select_by_threshold(x, cutoff, f.labels)
}
\description{
Select by thresholding

Select by thresholding
}
\keyword{internal}
