% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{splitDataIDX}
\alias{splitDataIDX}
\title{Split Data into Training and Test Sets}
\usage{
splitDataIDX(c_obs, v_obs, train_frac = 0.5)
}
\arguments{
\item{c_obs}{Integer vector of length \code{n}, giving the pre-labeled
partitioning variable status for each cell (may contain \code{NA}).}

\item{v_obs}{Integer vector of length \code{n}, giving the pre-labeled
infection status labels for each cell (may contain \code{NA}).}

\item{train_frac}{Numeric in \code{(0,1)}. Proportion of cells
within each variant (and missing group) assigned to the training
set. Defaults to \code{0.5}.}
}
\value{
A list with the following elements:
\describe{
\item{\code{generation_idx}}{Training indices.}
\item{\code{test_idx}}{Testing indices.}
}
}
\description{
Returns the indices to split the observed dataset into training
("generation") and test subsets, stratified by labels.
Cells with missing variant labels are also split proportionally.
}
