% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_set_inference.R
\name{thresholdScoring}
\alias{thresholdScoring}
\title{Score Cell Type and Viral Status Labels via Thresholding on Test Data}
\usage{
thresholdScoring(
  mod_results,
  testData,
  cutoffInfection = 0.9,
  cutoffPartition = "max",
  independentThres = TRUE
)
}
\arguments{
\item{mod_results}{A list containing model fitting results from \code{\link{fit_scDEcrypter}}.}

\item{testData}{A Seurat object or list containing test data.}

\item{cutoffInfection}{Numeric threshold in \code{[0,1]}, or \code{"max"}.}

\item{cutoffPartition}{Numeric threshold in \code{[0,1]}, or \code{"max"}.}

\item{independentThres}{If \code{TRUE}, marginalize over the other dimension.}
}
\value{
The input test data with predicted infection and partition states.
}
\description{
Assigns cell type and viral infection status labels to test cells based on
conditional weights.
}
