% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_m_step.R
\name{update_mu}
\alias{update_mu}
\title{Update Mean Parameters With Penalty}
\usage{
update_mu(Y, M, sigma2, W, lambda)
}
\arguments{
\item{Y}{Numeric matrix of observed data with dimensions
\code{(n x p)}, where \code{n} is the number of samples (cells)
and \code{p} is the number of features (e.g., genes).}

\item{M}{Numeric array of current mean estimates with dimensions
\code{(p x C x V)}, where \code{C} is the number of conditions
and \code{V} is the number of infection status}

\item{sigma2}{Numeric array of variance estimates with dimensions
\code{(p x C x V)}.}

\item{W}{Numeric 3D array of posterior weights (responsibilities)
with dimensions \code{(n x C x V)}.}

\item{lambda}{Numeric scalar; penalty parameter applied in the update.}
}
\value{
A numeric array of updated estimates.
}
\description{
Updates the mean parameter array \eqn{M} in the multiway mixture model
using penalized accelerated proximal gradient descent (APGD).
This version applies a row-wise penalty controlled by \code{lambda}.
}
