% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_funcs.R
\name{get_top_genes}
\alias{get_top_genes}
\title{Identify Top Variable Genes over Cell Partitions}
\usage{
get_top_genes(
  seurat_obj,
  partition_colname = "C.preLabel",
  n_genes_per_group = 50
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing single-cell RNA-seq data.}

\item{partition_colname}{Character. Name of the cell-level metadata column used to partition cells into groups (e.g., cell type, cluster).}

\item{n_genes_per_group}{Integer. Number of top variable genes to rank per group (default: 500).}
}
\value{
A character vector of unique gene names with highest variance across the specified groups.
}
\description{
Selects the most variable genes within each group of cells defined by a partitioning column in a Seurat object.
}
\details{
For each group defined by \code{partition_colname}, the function calculates the variance of each gene across cells in that group,
selects the top \code{n_genes_per_group} most variable genes, and then returns the unique set of the top \code{n_top} genes from each group.
}
