% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccPGD.Dm.R
\name{obj.func}
\alias{obj.func}
\title{Internal: Objective function}
\usage{
obj.func(M, a, S, D, lambda)
}
\arguments{
\item{M}{Numeric vector of current estimates.}

\item{a}{Numeric vector of observed values.}

\item{S}{Numeric weight vector.}

\item{D}{Difference matrix.}

\item{lambda}{Penalty parameter (numeric scalar).}
}
\value{
Numeric value of the objective function.
}
\description{
Computes the objective value for the proximal gradient descent algorithm.
This is an internal helper function and not intended to be called directly.
}
\keyword{internal}
