% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_funcs.R
\name{preprocess_scDEcrypter}
\alias{preprocess_scDEcrypter}
\title{Preprocess data for scDEcrypter}
\usage{
preprocess_scDEcrypter(
  Data,
  c_obs_var = "C.preLabel",
  v_obs_var = "V.preLabel",
  train_frac = 0.5,
  vs_method = c("shifted_log", "none")
)
}
\arguments{
\item{Data}{A Seurat object containing an RNA assay with raw counts data.}

\item{c_obs_var}{Character name of meta data column giving the pre-labeled
partitioning variable status for each cell (may contain \code{NA}).}

\item{v_obs_var}{Character name of meta data column giving the pre-labeled
infection status labels for each cell (may contain \code{NA}).}

\item{train_frac}{Numeric in \code{(0,1)}. Proportion of cells
within each variant (and missing group) assigned to the training
set. Defaults to \code{0.5}.}

\item{vs_method}{Transformation method passed to transformGamPoi. Default is \code{shifted_log}.}
}
\description{
Data will first be split into training and testing, followed
by independent variance stabilizing normalization.
}
