% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDEcrypter-package.R
\docType{package}
\name{scDEcrypter-package}
\alias{scDEcrypter}
\alias{scDEcrypter-package}
\title{scDEcrypter: An R package uses limited information to predict virus cell infection, and does differentially expressed gene expression analysis}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Luer Zhong \email{luerzhong@ufl.edu}

}
\keyword{internal}
