% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_set_helpers.R
\name{switchNames}
\alias{switchNames}
\title{Replace Dimension Names with Biological Labels}
\usage{
switchNames(inResults, infectionLabels, partitionLabels)
}
\description{
Replaces generic model dimension names with biological labels.
}
\keyword{internal}
